+++ node/Cargo.toml
@@ -12,7 +12,6 @@ targets = [ "x86_64-unknown-linux-gnu" ]
 
 [dependencies]
 clap = { features = [ "derive" ], workspace = true }
-hex-literal = { workspace = true }
 
 sc-cli = { workspace = true }
 sc-client-api = { workspace = true }
